fuzz修复

+++ test/fuzztest/addappInstallcontrolrule_fuzzer/addappInstallcontrolrule_fuzzer.cpp
@@ -30,7 +30,7 @@ namespace OHOS {
         std::vector<std::string> appIds;
         std::string appId (reinterpret_cast<const char*>(data), size);
         appIds.push_back(appId);
-        appControl.AddAppInstallControlRule(appIds, controlRuleType, reinterpret_cast<int32_t>(data));
+        appControl.AddAppInstallControlRule(appIds, controlRuleType, reinterpret_cast<uintptr_t>(data));
         return true;
     }
 }
