+++ test/fuzztest/deleteappinstallcontrolrule_fuzzer/deleteappinstallcontrolrule_fuzzer.cpp
@@ -30,8 +30,8 @@ namespace OHOS {
         std::vector<std::string> appIds;
         std::string appId (reinterpret_cast<const char*>(data), size);
         appIds.push_back(appId);
-        appControl.DeleteAppInstallControlRule(controlRuleType, appIds, reinterpret_cast<int32_t>(data));
-        appControl.DeleteAppInstallControlRule(controlRuleType, reinterpret_cast<int32_t>(data));
+        appControl.DeleteAppInstallControlRule(controlRuleType, appIds, reinterpret_cast<uintptr_t>(data));
+        appControl.DeleteAppInstallControlRule(controlRuleType, reinterpret_cast<uintptr_t>(data));
         return true;
     }
 }
